requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pandas==2.2.3
matplotlib==3.10.1
reportlab==4.2.5
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.1.0
chardet==5.2.0
